rename ardcvocabs module to generic researchvocabs module
