--- ardcvocabs/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-au.org.aodn.ardcvocabs.configuration.ArdcAutoConfiguration
