+++ researchvocabs/pom.xml
@@ -10,7 +10,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>ardcvocabs</artifactId>
+    <artifactId>researchvocabs</artifactId>
     <packaging>jar</packaging>
 
     <properties>
