+++ researchvocabs/src/test/resources/databag/vocab0.json
