+++ researchvocabs/src/test/resources/databag/vocab3.json
