+++ researchvocabs/src/test/resources/databag/vocab_discovery0.json
